move DcaUtil
