+++ tests/Util/UtilsTest.php
@@ -12,7 +12,7 @@ use Contao\TestCase\ContaoTestCase;
 use HeimrichHannot\UtilsBundle\Util\ContainerUtil;
 use HeimrichHannot\UtilsBundle\Util\AnonymizeUtil;
 use HeimrichHannot\UtilsBundle\Util\DatabaseUtil\DatabaseUtil;
-use HeimrichHannot\UtilsBundle\Util\Dca\DcaUtil;
+use HeimrichHannot\UtilsBundle\Util\DcaUtil\DcaUtil;
 use HeimrichHannot\UtilsBundle\Util\Html\HtmlUtil;
 use HeimrichHannot\UtilsBundle\Util\Locale\LocaleUtil;
 use HeimrichHannot\UtilsBundle\Util\Model\ModelUtil;
